refactor: 업체 dto 폴더 이동
#57
