+++ com.bulbas23r.client.company/src/main/java/com/bulbas23r/client/company/application/service/CompanyService.java
@@ -1,8 +1,8 @@
 package com.bulbas23r.client.company.application.service;
 
-import com.bulbas23r.client.company.application.dto.CompanyCreateRequestDto;
-import com.bulbas23r.client.company.application.dto.CompanyResponseDto;
-import com.bulbas23r.client.company.application.dto.CompanyUpdateRequestDto;
+import com.bulbas23r.client.company.presentation.dto.CompanyCreateRequestDto;
+import com.bulbas23r.client.company.presentation.dto.CompanyResponseDto;
+import com.bulbas23r.client.company.presentation.dto.CompanyUpdateRequestDto;
 import com.bulbas23r.client.company.domain.model.CompanyType;
 import common.utils.PageUtils.CommonSortBy;
 import java.util.UUID;
