+++ com.bulbas23r.client.company/src/main/java/com/bulbas23r/client/company/infrastructure/messaging/CompanyEventProducer.java
@@ -7,7 +7,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Service;
 
-@Slf4j
 @Service
 @RequiredArgsConstructor
 public class CompanyEventProducer {
