+++ com.bulbas23r.client.company/src/main/java/com/bulbas23r/client/company/infrastructure/persistence/CompanyQueryRepositoryImpl.java
@@ -1,7 +1,7 @@
 package com.bulbas23r.client.company.infrastructure.persistence;
 
 import ch.qos.logback.core.util.StringUtil;
-import com.bulbas23r.client.company.application.dto.CompanyResponseDto;
+import com.bulbas23r.client.company.presentation.dto.CompanyResponseDto;
 import com.bulbas23r.client.company.domain.model.CompanyType;
 import com.bulbas23r.client.company.domain.model.QCompany;
 import com.bulbas23r.client.company.domain.repository.CompanyQueryRepository;
